refactor: 프로필 관련 컴포넌트 리팩토링 (#290)
* refactor: vscode 경고 수정 (#288)

- vscode에서 경고로 표시되는 부분 수정

* refactor: mypage profile mock-date 사용 (#288)

- USE_MOCK = true로 변경하여 api 대신 목데이터 사용

* refactor: ProfileBackgroundImage 파일 이름 변경 (#288)

- 파일 이름 변경 및 주석 수정

* refactor: ProfileBaseImage 컴포넌트로 이름 변경 (#288)

- ProfileBaseUi -> ProfileBaseImage 로 컴포넌트 이름 변경

* refactor: ProfileGenresType 컴포넌트 생성 (#288)

- 기존 ProfileGenresUi -> ProfileGenresType 으로 컴포넌트 이름 변경
- deviceType에 따른 UI 렌더링 분기

* refactor: ProfileGlowEvent 컴포넌트 생성 (#288)

- 기존 ProfileGlowUi -> ProfileGlowEvent 로 컴포넌트 이름 변경
- deviceType에 따른 UI 렌더링 분기

* refactor: ProfileHoverUi 삭제 및 컴포넌트 분리 (#288)

- 기존 ProfileHoverUi 컴포넌트 삭제 후 delete, edit 관련 컴포넌트로 분리

* refactor: ProfileName 컴포넌트 생성 (#288)

- 기존 ProfileNameUi -> ProfileName 컴포넌트로 이름 변경
- deviceType에 따른 UI 렌더링 분기

* refactor: 프로필 관련 컴포넌트 리팩토링 (#288)

- deviceType (mobile, desktop) 타입 생성
- deviceType에 따른 프로필 컴포넌트 분기 처리
- 불필요한 코드 및 주석 제거
