+++ src/components/feature/mypage/index.ts
@@ -2,20 +2,6 @@ export { default as Profile } from './Profile'
 export { default as DashBoard } from './DashBoard'
 export { default as WishList } from './WishList'
 export { default as MyReviewList } from './MyReviewList'
-// ui
-export { ProfileDesktopUi, ProfileMobileUi } from './ui/profile-ui/ProfileUi'
-export {
-  DeleteProfileImageUi,
-  EditProfileImageUi,
-} from './ui/profile-ui/ProfileHoverUi'
-export {
-  ProfileGenresDesktopUi,
-  ProfileGenresMobileUi,
-} from './ui/profile-ui/ProfileGenresUi'
-export {
-  ProfileDesktopNameUi,
-  ProfileMobileNameUi,
-} from './ui/profile-ui/ProfileNameUi'
 export { default as DashBoardCardUi } from './ui/DashBoardCardUi'
 export { default as MyReviewListSectionUi } from './ui/MyReviewListSectionUi'
 export { default as MyReviewListWelcomeUi } from './ui/MyReviewListWelcomeUi'
@@ -24,6 +10,5 @@ export { default as WishListCarouselNavUi } from './ui/WishListCarouselNavUi'
 export { default as WishListCarouselUi } from './ui/WishListCarouselUi'
 export { default as WishListEmptyUi } from './ui/WishListEmptyUi'
 export { default as WishListTitleUi } from './ui/WishListTitleUi'
-// user-info
 export { default as InfoBox } from './user-info/user-delete/InfoBox'
 export { default as UserDeleteClient } from './user-info/user-delete/UserDeleteClient'
