+++ src/components/feature/mypage/ui/profile-ui/EditProfileImage.tsx
@@ -1,18 +1,17 @@
 'use client'
 
 import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { Pencil, X } from 'lucide-react'
+import { Pencil } from 'lucide-react'
 import { ChangeEvent, useRef, useState } from 'react'
 
 import { postProfileApi } from '@/api/fetchers/profileFetchers'
-import useDeleteProfileImage from '@/api/queries/useDeleteProfileImage'
 import { IMAGE_ERROR_MESSAGES, IMAGE_UPLOAD_CONFIG } from '@/constants'
 import useToast from '@/hooks/useToast'
 import { cn, validateFileSize, validateFileType } from '@/utils'
 
 import ProfileImageCropDialog from './ProfileImageCropDialogUi'
 
-function EditProfileImageUi() {
+export default function EditProfileImageUi() {
   const [isDialogOpen, setIsDialogOpen] = useState(false)
   const [selectedImage, setSelectedImage] = useState<string | null>(null)
   const fileInputRef = useRef<HTMLInputElement>(null)
@@ -114,38 +113,3 @@ function EditProfileImageUi() {
     </>
   )
 }
-
-function DeleteProfileImageUi() {
-  const { mutate: deleteImage } = useDeleteProfileImage()
-
-  const handleDelete = (e: React.MouseEvent) => {
-    e.stopPropagation()
-
-    // 삭제 확인
-    const confirmed = window.confirm('프로필 이미지를 삭제하시겠습니까?')
-
-    if (!confirmed) {
-      return
-    }
-
-    deleteImage()
-  }
-
-  return (
-    <button
-      title="삭제 버튼"
-      className={cn(
-        'absolute top-2.5 left-2.5 z-6 flex h-35.75 w-35.75 items-center justify-center rounded-full',
-        'border border-white/10 bg-white/5 backdrop-blur-md',
-        'opacity-0 transition-all duration-300',
-        'group-hover:scale-105 group-hover:border-white/20 group-hover:bg-white/10 group-hover:opacity-100',
-        'group-hover:shadow-[0_4px_12px_rgba(168,85,247,0.2)]'
-      )}
-      onClick={handleDelete}
-    >
-      <X className="size-8 text-text-light transition-colors duration-300 group-hover:text-text-light" />
-    </button>
-  )
-}
-
-export { DeleteProfileImageUi, EditProfileImageUi }
