+++ src/components/feature/mypage/ui/profile-ui/ProfileBackgroundImage.tsx
@@ -3,16 +3,18 @@ import Image from 'next/image'
 import profileBackgroundDesktop from '@/assets/images/profile/profile-background-desktop.svg'
 import profileBackgroundMobile from '@/assets/images/profile/profile-background-mobile.svg'
 
-export default function ProfileBackgroundUi() {
+/**
+ * 프로필 배경 이미지 컴포넌트
+ * css에 따라 모바일과 데스크탑 이미지가 분기되어 렌더링
+ */
+export default function ProfileBackgroundImage() {
   return (
     <>
-      {/* 프로필 / 배경 / 모바일 */}
       <Image
         src={profileBackgroundMobile}
         alt="profile background mobile"
         className="absolute top-0 left-0 z-1 rounded-full lg:h-auto lg:w-auto"
       />
-      {/* 프로필 / 배경 / 데스크탑 */}
       <Image
         src={profileBackgroundDesktop}
         alt="profile background desktop"
