+++ src/components/feature/mypage/ui/profile-ui/ProfileImageUi.tsx
@@ -3,7 +3,9 @@
 import { PROFILE_TEXT } from '@/constants'
 import { cn, getFullImageUrl } from '@/utils'
 
-import { DeleteProfileImageUi, EditProfileImageUi, ProfileBaseUi } from '.'
+import DeleteProfileImageUi from './DeleteProfileImage'
+import EditProfileImageUi from './EditProfileImage'
+import ProfileBaseImage from './ProfileBaseImage'
 
 type ProfileImageUiProps = {
   imageUrl: string
@@ -14,13 +16,13 @@ export default function ProfileImageUi({ imageUrl }: ProfileImageUiProps) {
 
   return (
     <>
-      {/* 회전하는 그라데이션 링 - 천천히 (8초) ⭐ */}
+      {/** 회전하는 그라데이션 링 - 천천히 (8초) */}
       <div className="absolute top-2.5 left-2.5 z-3 h-35.75 w-35.75 animate-spin-slow">
         <div className="h-full w-full rounded-full bg-linear-to-r from-main-purple via-main-fuchsia to-main-violet p-0.5 opacity-40 transition-opacity duration-500 group-hover:opacity-70">
           <div className="h-full w-full rounded-full bg-transparent" />
         </div>
       </div>
-      {/* 프로필 / 이미지 / 회원 - 회전 없음 ⭐ */}
+      {/** 프로필 / 이미지 / 회원 - 회전 없음 */}
       {fullImageUrl && fullImageUrl.length > 0 ? (
         <>
           {/* eslint-disable-next-line @next/next/no-img-element */}
@@ -43,7 +45,7 @@ export default function ProfileImageUi({ imageUrl }: ProfileImageUiProps) {
       ) : (
         <EditProfileImageUi />
       )}
-      <ProfileBaseUi />
+      <ProfileBaseImage />
     </>
   )
 }
