+++ src/components/feature/mypage/ui/profile-ui/ProfileUi.tsx
@@ -2,63 +2,25 @@
 
 import { useGetPrefernece } from '@/api/queries/useGetUserPreference'
 import { GetUserMe } from '@/types/api-response/user-response'
+import { DeviceType } from '@/types/deviceType'
 
-import {
-  ProfileBackgroundUi,
-  ProfileDesktopGlowUi,
-  ProfileDesktopNameUi,
-  ProfileGenresDesktopUi,
-  ProfileGenresMobileUi,
-  ProfileImageUi,
-  ProfileMobileGlowUi,
-  ProfileMobileNameUi,
-} from '.'
+import ProfileBackgroundImage from './ProfileBackgroundImage'
+import ProfileGenresType from './ProfileGenresType'
+import ProfileGlowEvent from './ProfileGlowEvent'
+import ProfileImageUi from './ProfileImageUi'
+import ProfileName from './ProfileName'
 
 type ProfileUiProps = {
   nickname: GetUserMe['nickname']
   imageUrl: string
+  deviceType: DeviceType
 }
 
-function ProfileDesktopUi({ imageUrl, nickname }: ProfileUiProps) {
-  const { data } = useGetPrefernece()
-
-  const mergedPreferences = [
-    ...(data?.genres ?? []).map((g) => ({
-      id: g.id,
-      name: g.genre,
-    })),
-    ...(data?.tags ?? []).map((t) => ({
-      id: t.id,
-      name: t.tag,
-    })),
-  ]
-  return (
-    <div className="relative">
-      <ProfileDesktopGlowUi />
-      <div className="group relative flex h-auto w-auto cursor-pointer flex-col items-center justify-center">
-        <ProfileImageUi imageUrl={imageUrl} />
-      </div>
-      <ProfileDesktopNameUi nickname={nickname} />
-      <ProfileGenresDesktopUi
-        genre="GENRES (DESKTOP)"
-        variant="main"
-        className="absolute top-15.25 left-38 z-2"
-      />
-      <ProfileBackgroundUi />
-      <div className="absolute top-30 left-41 z-2 flex flex-wrap items-center gap-2">
-        {mergedPreferences.map((items) => (
-          <ProfileGenresDesktopUi
-            key={`pref-${items.id}`}
-            genre={items.name}
-            className="px-4 py-1.5 text-sm"
-          />
-        ))}
-      </div>
-    </div>
-  )
-}
-
-function ProfileMobileUi({ imageUrl, nickname }: ProfileUiProps) {
+export default function ProfileUi({
+  imageUrl,
+  nickname,
+  deviceType,
+}: ProfileUiProps) {
   const { data } = useGetPrefernece()
 
   const mergedPreferences = [
@@ -72,32 +34,61 @@ function ProfileMobileUi({ imageUrl, nickname }: ProfileUiProps) {
     })),
   ]
 
-  return (
-    <div className="relative flex flex-col items-center">
-      <ProfileMobileGlowUi />
-      <div className="relative flex h-40.75 w-40.75 flex-col items-center justify-center">
-        <div className="group relative z-10 h-40.75 w-40.75 cursor-pointer">
+  if (deviceType === 'desktop') {
+    return (
+      <div className="relative">
+        <ProfileGlowEvent deviceType="desktop" />
+        <div className="group relative flex h-auto w-auto cursor-pointer flex-col items-center justify-center">
           <ProfileImageUi imageUrl={imageUrl} />
         </div>
-        <ProfileBackgroundUi />
+        <ProfileName nickname={nickname} deviceType="desktop" />
+        <ProfileGenresType
+          deviceType="desktop"
+          genre="GENRES (DESKTOP)"
+          variant="main"
+          className="absolute top-15.25 left-38 z-2"
+        />
+        <ProfileBackgroundImage />
+        <div className="absolute top-30 left-41 z-2 flex flex-wrap items-center gap-2">
+          {mergedPreferences.map((items) => (
+            <ProfileGenresType
+              deviceType="desktop"
+              key={`pref-${items.id}`}
+              genre={items.name}
+              className="px-4 py-1.5 text-sm"
+            />
+          ))}
+        </div>
       </div>
-      <ProfileMobileNameUi nickname={nickname} />
-      <ProfileGenresMobileUi
-        genre="GENRES (MOBILE)"
-        variant="main"
-        className="z-7 mt-5 block px-4 py-2 text-sm"
-      />
-      <div className="z-7 mt-5 flex w-[70vw] flex-wrap items-center justify-center gap-2">
-        {mergedPreferences.map((items) => (
-          <ProfileGenresMobileUi
-            key={`pref-${items.id}`}
-            genre={items.name}
-            className="px-3 py-1.5 text-xs"
-          />
-        ))}
+    )
+  } else {
+    return (
+      <div className="relative flex flex-col items-center">
+        <ProfileGlowEvent deviceType="mobile" />
+        <div className="relative flex h-40.75 w-40.75 flex-col items-center justify-center">
+          <div className="group relative z-10 h-40.75 w-40.75 cursor-pointer">
+            <ProfileImageUi imageUrl={imageUrl} />
+          </div>
+          <ProfileBackgroundImage />
+        </div>
+        <ProfileName nickname={nickname} deviceType="mobile" />
+        <ProfileGenresType
+          deviceType="mobile"
+          genre="GENRES (MOBILE)"
+          variant="main"
+          className="z-7 mt-5 block px-4 py-2 text-sm"
+        />
+        <div className="z-7 mt-5 flex w-[70vw] flex-wrap items-center justify-center gap-2">
+          {mergedPreferences.map((items) => (
+            <ProfileGenresType
+              deviceType="mobile"
+              key={`pref-${items.id}`}
+              genre={items.name}
+              className="px-3 py-1.5 text-xs"
+            />
+          ))}
+        </div>
       </div>
-    </div>
-  )
+    )
+  }
 }
-
-export { ProfileDesktopUi, ProfileMobileUi }
