+++ src/types/deviceType.ts
@@ -1,0 +1,1 @@
+export type DeviceType = 'desktop' | 'mobile'
